+++ domain/src/main/java/com/stansdevhouse/domain/model/Game.kt
@@ -1,3 +1,3 @@
 package com.stansdevhouse.domain.model
 
-data class Game(val title: String, val imageUrl: String, val id: Int)
+data class Game(val title: String, val imageUrl: String, val id: Long)
